pyproject.toml
[project]
dynamic = ["version"]
name = "novi"
description = "A simple, fast, and lightweight YouTube downloader."
readme = "README.md"
dependencies = [
    "yt-dlp",
    "webvtt-py",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["novi"]
